Move PackageReport serializer to serializers.py
Move PackageReport serializer to serializers.py
