+++ django/thunderstore/community/api/experimental/serializers.py
@@ -1,9 +1,11 @@
 from rest_framework import serializers
+from rest_framework.relations import PrimaryKeyRelatedField
 
 from thunderstore.community.models import PackageCategory
 from thunderstore.repository.api.experimental.serializers import (
     CommunityFilteredModelChoiceField,
 )
+from thunderstore.repository.models import PackageVersion
 
 
 class PackageListingUpdateRequestSerializer(serializers.Serializer):
@@ -24,4 +26,22 @@ class PackageCategoryExperimentalSerializer(serializers.Serializer):
 class PackageListingUpdateResponseSerializer(serializers.Serializer):
     categories = serializers.ListSerializer(
         child=PackageCategoryExperimentalSerializer()
+    )
+
+
+class PackageListingReportRequestSerializer(serializers.Serializer):
+    version = PrimaryKeyRelatedField(
+        required=False,
+        allow_null=True,
+        queryset=PackageVersion.objects.all(),
+    )
+    reason = serializers.CharField(
+        required=True,
+        allow_blank=False,
+        allow_null=False,
+    )
+    description = serializers.CharField(
+        required=False,
+        allow_blank=True,
+        allow_null=True,
     )
